+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 In next release ...
 
+- The [ts-typed-events](https://www.npmjs.com/package/ts-typed-events)
+  dependency was updated to version 3.0.0.
+
 1.2.1 (2021-12-23)
 ------------------